Remove Wallhaven provider
It is currently completely defunct. I'm not sure if that is because the Wallhaven API changed, its URLs changed, because of the GTK4 update, or because of something on my own computer blocking it.
I'm removing it regardless. I'm not interested in fixing this, but if someone wants to get it back up and running they are welcome to try.