Fix handling of names of providers
We don't have the __name__ property any more. I tried providing a name via the constructor argument of _Provider but that doesn't work because it's a class and not an instance. So this is the solution now, just a property of the class. Not enforced via the parent class, but oh well.