Change dark mode wallpaper too
This extension chooses a random wallpaper from a certain directory. So it doesn't make sense to have different wallpapers for dark mode then. You might want to have a different directory for dark mode, but for now you'll just have to change the directory manually when you switch. The light mode and dark mode wallpapers will be the same.

Fixes #8.